Driver list is running

+++ BlazorApp1/Models/Driver.cs
@@ -6,23 +6,23 @@ namespace BlazorApp1.Models
     public class Driver
     {
         [Key]
-        public Int32 idDriver { get; set; }
+        public int idDriver { get; set; }
         public String firstName { get; set; }
         public String lastName { get; set; }
         public String phoneNumber { get; set; }
         public String email { get; set; }
         public String driverLicense { get; set; }
         public String address { get; set; }
-        public Int32 idCity { get; set; }
-        public Int32 idState { get; set; }
+        public int idCity { get; set; }
+        public int idState { get; set; }
         public String zipcode { get; set; }
-        public Int32 idCountryFrom { get; set; }
-        public Int32 idVehicleType { get; set; }
+        public int idCountryFrom { get; set; }
+        public int idVehicleType { get; set; }
         public String vehicleModel { get; set; }
         public String vehicleBrand { get; set; }
         public String vehicleVIN { get; set; }
-        public Int32 vehicleYear { get; set; }
+        public Int16 vehicleYear { get; set; }
         public String vehicleColor { get; set; }
-        public Int32 idDriverStatus { get; set; }
+        public int idDriverStatus { get; set; }
     }
 }
